Pagina 15 En errors

+++ src/main/java/edu/alumno/hector/api_rest_mysql_futbol/security/repository/UsuarioRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import edu.alumno.hector.api_rest_mysql_futbol.security.entity.UsuarioDb;
 
-public interface UsuarioRepository extends JpaRepository<UsuarioDb, Integer>{
+public interface UsuarioRepository extends JpaRepository<UsuarioDb, Long>{
     Optional<UsuarioDb> findByNickname(String nickname);
     boolean existsByNickname(String nickname);
     boolean existsByEmail(String email);
